WEB-4460 migrate tests for async actions
