+++ __tests__/unit/app/redux/actions/async.test.js
@@ -2,28 +2,17 @@
 /* global describe */
 /* global it */
 /* global expect */
-/* global beforeEach */
-/* global before */
-/* global afterEach */
-/* global after */
-/* global context */
 
 import configureStore from 'redux-mock-store';
 import { thunk } from 'redux-thunk';
 import trackingMiddleware from '../../../../../app/redux/utils/trackingMiddleware';
-import moment from 'moment';
-import _ from 'lodash';
-
-import isTSA from 'tidepool-standard-action';
 
 import initialState from '../../../../../app/redux/reducers/initialState';
 
 import * as ErrorMessages from '../../../../../app/redux/constants/errorMessages';
-import * as UserMessages from '../../../../../app/redux/constants/usrMessages';
 
 // need to require() async in order to rewire utils inside
 const async = require('../../../../../app/redux/actions/async');
-const sync = require('../../../../../app/redux/actions/sync');
 
 describe('Actions', () => {
   const trackMetric = jest.fn();
@@ -48,9 +37,6 @@ describe('Actions', () => {
           { type: 'CREATE_CLINIC_SITE_REQUEST' },
           { type: 'CREATE_CLINIC_SITE_SUCCESS', payload: { clinicId, site: newSite } },
         ];
-        _.each(expectedActions, (action) => {
-          expect(isTSA(action)).toBe(true);
-        });
 
         let store = mockStore({ blip: initialState });
         store.dispatch(async.createClinicSite(api, clinicId, { name: 'Site Alpha' }));
@@ -76,15 +62,175 @@ describe('Actions', () => {
           { type: 'CREATE_CLINIC_SITE_REQUEST' },
           { type: 'CREATE_CLINIC_SITE_FAILURE', error: err, meta: { apiError: { status: 500, body: 'Error!' } } },
         ];
-        _.each(expectedActions, (action) => {
-          expect(isTSA(action)).toBe(true);
-        });
+
         let store = mockStore({ blip: initialState });
         store.dispatch(async.createClinicSite(api, clinicId, { name: 'Site Charlie' }));
 
         const actions = store.getActions();
         expect(actions).toStrictEqual(expectedActions);
         expect(api.clinics.createClinicSite).toHaveBeenCalledTimes(1);
+      });
+    });
+
+    describe('selectClinic', () => {
+      const clinicId = 'clinic123';
+
+      const expectedUIDetails = {
+        entitlements: {
+          patientTags: false,
+          clinicSites: false,
+          prescriptions: false,
+          rpmReport: false,
+          summaryDashboard: false,
+          tideDashboard: false,
+        },
+        patientLimitEnforced: false,
+        planName: 'internationalBase',
+        ui: {
+          display: {
+            patientCount: true,
+            patientLimit: false,
+            planName: false,
+            workspacePlan: false,
+            workspaceLimitDescription: false,
+            workspaceLimitFeedback: false,
+            workspaceLimitResolutionLink: false,
+          },
+          text: {
+            planDisplayName: 'Base',
+            limitDescription: undefined,
+            limitFeedback: undefined,
+            limitResolutionLink: undefined,
+          },
+          warnings: {
+            limitApproaching: false,
+            limitReached: false,
+          },
+        },
+      };
+
+      it('should trigger SELECT_CLINIC_SUCCESS, FETCH_CLINIC_PATIENT_COUNTS_SUCCESS, and FETCH_CLINIC_PATIENT_COUNT_SETTINGS_SUCCESS for a successful request', () => {
+        const countResults = { plan: 33 };
+        const settingsResults = { bar: 'baz' };
+
+        let api = {
+          clinics: {
+            getClinicPatientCount: jest.fn().mockImplementation((_arg1, cb) => cb(null, countResults)),
+            getClinicPatientCountSettings: jest.fn().mockImplementation((_arg1, cb) => cb(null, settingsResults)),
+          },
+        };
+
+        let expectedActions = [
+          { type: 'SELECT_CLINIC_SUCCESS', payload: { clinicId } },
+          { type: 'tideDashboardFilters/setTideDashboardFilters', payload: undefined },
+          { type: 'FETCH_CLINIC_PATIENT_COUNTS_REQUEST' },
+          { type: 'FETCH_CLINIC_PATIENT_COUNT_SETTINGS_REQUEST' },
+          { type: 'FETCH_CLINIC_PATIENT_COUNTS_SUCCESS', payload: { clinicId, patientCounts: countResults } },
+          { type: 'FETCH_CLINIC_PATIENT_COUNT_SETTINGS_SUCCESS', payload: { clinicId, patientCountSettings: settingsResults } },
+          { type: 'SET_CLINIC_UI_DETAILS', payload: { clinicId, uiDetails: expectedUIDetails } },
+        ];
+
+        let store = mockStore({ blip: {
+          ...initialState,
+          clinics: {
+            [clinicId]: {
+              patientCounts: undefined,
+              patientCountSettings: undefined,
+            },
+          },
+        } });
+
+        store.dispatch(async.selectClinic(api, clinicId));
+
+        const actions = store.getActions();
+        expect(actions).toStrictEqual(expectedActions);
+        expect(api.clinics.getClinicPatientCount).toHaveBeenCalledWith(clinicId, expect.any(Function));
+        expect(api.clinics.getClinicPatientCount).toHaveBeenCalledTimes(1);
+        expect(api.clinics.getClinicPatientCountSettings).toHaveBeenCalledWith(clinicId, expect.any(Function));
+        expect(api.clinics.getClinicPatientCountSettings).toHaveBeenCalledTimes(1);
+      });
+
+      it('should trigger SELECT_CLINIC_SUCCESS, but not FETCH_CLINIC_PATIENT_COUNTS_REQUEST or FETCH_CLINIC_PATIENT_COUNT_SETTINGS_REQUEST for a successful request if data available in clinic state', () => {
+        let api = {
+          clinics: {
+            getClinicPatientCount: jest.fn(),
+            getClinicPatientCountSettings: jest.fn(),
+          },
+        };
+
+        let expectedActions = [
+          { type: 'SELECT_CLINIC_SUCCESS', payload: { clinicId } },
+          { type: 'tideDashboardFilters/setTideDashboardFilters', payload: undefined },
+          { type: 'SET_CLINIC_UI_DETAILS', payload: { clinicId, uiDetails: expectedUIDetails } },
+        ];
+
+        let store = mockStore({ blip: {
+          ...initialState,
+          clinics: {
+            [clinicId]: {
+              patientCounts: { plan: 33 },
+              patientCountSettings: { foo: 'bar' },
+            },
+          },
+        } });
+
+        store.dispatch(async.selectClinic(api, clinicId));
+
+        const actions = store.getActions();
+        expect(actions).toStrictEqual(expectedActions);
+        expect(api.clinics.getClinicPatientCount).not.toHaveBeenCalled();
+        expect(api.clinics.getClinicPatientCountSettings).not.toHaveBeenCalled();
+      });
+
+      it('should trigger FETCH_CLINIC_PATIENT_COUNTS_FAILURE and FETCH_CLINIC_PATIENT_COUNT_SETTINGS_FAILURE and it should call error once for a failed request', () => {
+        let api = {
+          clinics: {
+            getClinicPatientCount: jest.fn().mockImplementation((_arg1, cb) => cb({ status: 500, body: 'Count Error!' }, null)),
+            getClinicPatientCountSettings: jest.fn().mockImplementation((_arg1, cb) => cb({ status: 500, body: 'Settings Error!' }, null)),
+          },
+        };
+
+        let countErr = new Error(ErrorMessages.ERR_FETCHING_CLINIC_PATIENT_COUNTS);
+        countErr.status = 500;
+
+        let settingsErr = new Error(ErrorMessages.ERR_FETCHING_CLINIC_PATIENT_COUNT_SETTINGS);
+        settingsErr.status = 500;
+
+        let expectedActions = [
+          { type: 'SELECT_CLINIC_SUCCESS', payload: { clinicId } },
+          { type: 'tideDashboardFilters/setTideDashboardFilters', payload: undefined },
+          { type: 'FETCH_CLINIC_PATIENT_COUNTS_REQUEST' },
+          { type: 'FETCH_CLINIC_PATIENT_COUNT_SETTINGS_REQUEST' },
+          {
+            type: 'FETCH_CLINIC_PATIENT_COUNTS_FAILURE',
+            error: countErr,
+            meta: { apiError: { status: 500, body: 'Count Error!' } },
+          },
+          {
+            type: 'FETCH_CLINIC_PATIENT_COUNT_SETTINGS_FAILURE',
+            error: settingsErr,
+            meta: { apiError: { status: 500, body: 'Settings Error!' } },
+          },
+        ];
+
+        let store = mockStore({ blip: {
+          ...initialState,
+          clinics: {
+            [clinicId]: {
+              patientCounts: undefined,
+              patientCountSettings: undefined,
+            },
+          },
+        } });
+
+        store.dispatch(async.selectClinic(api, clinicId));
+
+        const actions = store.getActions();
+        expect(actions).toStrictEqual(expectedActions);
+        expect(api.clinics.getClinicPatientCount).toHaveBeenCalledWith(clinicId, expect.any(Function));
+        expect(api.clinics.getClinicPatientCount).toHaveBeenCalledTimes(1);
+        expect(api.clinics.getClinicPatientCountSettings).toHaveBeenCalledWith(clinicId, expect.any(Function));
+        expect(api.clinics.getClinicPatientCountSettings).toHaveBeenCalledTimes(1);
       });
     });
   });
