+++ test/unit/redux/actions/async.test.js
@@ -9445,249 +9445,6 @@ describe('Actions', () => {
       });
     });
 
-    describe('selectClinic', () => {
-      it('should trigger SELECT_CLINIC_SUCCESS, FETCH_CLINIC_PATIENT_COUNTS_SUCCESS, and FETCH_CLINIC_PATIENT_COUNT_SETTINGS_SUCCESS for a successful request', () => {
-        const clinicId = 'clinic123';
-        const countResults = { plan: 33 };
-        const settingsResults = {bar: 'baz'};
-
-        let api = {
-          clinics: {
-            getClinicPatientCount: sinon
-              .stub()
-              .callsArgWith(1, null, countResults),
-            getClinicPatientCountSettings: sinon
-              .stub()
-              .callsArgWith(1, null, settingsResults),
-          },
-        };
-
-        let expectedActions = [
-          { type: 'SELECT_CLINIC_SUCCESS', payload: { clinicId } },
-          { type: 'tideDashboardFilters/setTideDashboardFilters', payload: undefined },
-          { type: 'FETCH_CLINIC_PATIENT_COUNTS_REQUEST' },
-          { type: 'FETCH_CLINIC_PATIENT_COUNT_SETTINGS_REQUEST' },
-          {
-            type: 'FETCH_CLINIC_PATIENT_COUNTS_SUCCESS',
-            payload: { clinicId, patientCounts: countResults },
-          },
-          {
-            type: 'FETCH_CLINIC_PATIENT_COUNT_SETTINGS_SUCCESS',
-            payload: { clinicId, patientCountSettings: settingsResults },
-          },
-          {
-            type: 'SET_CLINIC_UI_DETAILS',
-            payload: {
-              clinicId: 'clinic123',
-              uiDetails: {
-                entitlements: {
-                  patientTags: false,
-                  clinicSites: false,
-                  prescriptions: false,
-                  rpmReport: false,
-                  summaryDashboard: false,
-                  tideDashboard: false,
-                },
-                patientLimitEnforced: false,
-                planName: 'internationalBase',
-                ui: {
-                  display: {
-                    patientCount: true,
-                    patientLimit: false,
-                    planName: false,
-                    workspacePlan: false,
-                    workspaceLimitDescription: false,
-                    workspaceLimitFeedback: false,
-                    workspaceLimitResolutionLink: false,
-                  },
-                  text: {
-                    planDisplayName: 'Base',
-                    limitDescription: undefined,
-                    limitFeedback: undefined,
-                    limitResolutionLink: undefined,
-                  },
-                  warnings: {
-                    limitApproaching: false,
-                    limitReached: false,
-                  },
-                },
-              },
-            },
-          },
-        ];
-        _.each(expectedActions, (action) => {
-          expect(isTSA(action)).to.be.true;
-        });
-
-        let store = mockStore({ blip: {
-          ...initialState,
-          clinics: {
-            [clinicId]: {
-              patientCounts: undefined,
-              patientCountSettings: undefined,
-            },
-          },
-        } });
-
-        store.dispatch(
-          async.selectClinic(api, clinicId)
-        );
-
-        const actions = store.getActions();
-        expect(actions).to.eql(expectedActions);
-        expect(api.clinics.getClinicPatientCount.callCount).to.equal(1);
-        expect(api.clinics.getClinicPatientCountSettings.callCount).to.equal(1);
-      });
-
-      it('should trigger SELECT_CLINIC_SUCCESS, but not FETCH_CLINIC_PATIENT_COUNTS_REQUEST or FETCH_CLINIC_PATIENT_COUNT_SETTINGS_REQUEST for a successful request if data available in clinic state', () => {
-        const clinicId = 'clinic123';
-
-        let api = {
-          clinics: {
-            getClinicPatientCount: sinon.stub(),
-            getClinicPatientCountSettings: sinon.stub(),
-          },
-        };
-
-        let expectedActions = [
-          { type: 'SELECT_CLINIC_SUCCESS', payload: { clinicId } },
-          { type: 'tideDashboardFilters/setTideDashboardFilters', payload: undefined },
-          {
-            type: 'SET_CLINIC_UI_DETAILS',
-            payload: {
-              clinicId: 'clinic123',
-              uiDetails: {
-                entitlements: {
-                  patientTags: false,
-                  clinicSites: false,
-                  prescriptions: false,
-                  rpmReport: false,
-                  summaryDashboard: false,
-                  tideDashboard: false,
-                },
-                patientLimitEnforced: false,
-                planName: 'internationalBase',
-                ui: {
-                  display: {
-                    patientCount: true,
-                    patientLimit: false,
-                    planName: false,
-                    workspacePlan: false,
-                    workspaceLimitDescription: false,
-                    workspaceLimitFeedback: false,
-                    workspaceLimitResolutionLink: false,
-                  },
-                  text: {
-                    planDisplayName: 'Base',
-                    limitDescription: undefined,
-                    limitFeedback: undefined,
-                    limitResolutionLink: undefined,
-                  },
-                  warnings: {
-                    limitApproaching: false,
-                    limitReached: false,
-                  },
-                },
-              },
-            },
-          },
-        ];
-        _.each(expectedActions, (action) => {
-          expect(isTSA(action)).to.be.true;
-        });
-
-        let store = mockStore({ blip: {
-          ...initialState,
-          clinics: {
-            [clinicId]: {
-              patientCounts: { plan: 33 },
-              patientCountSettings: { foo: 'bar' },
-            },
-          },
-        } });
-
-        store.dispatch(
-          async.selectClinic(api, clinicId)
-        );
-
-        const actions = store.getActions();
-        expect(actions).to.eql(expectedActions);
-        expect(api.clinics.getClinicPatientCount.callCount).to.equal(0);
-        expect(api.clinics.getClinicPatientCountSettings.callCount).to.equal(0);
-      });
-
-      it('should trigger FETCH_CLINIC_PATIENT_COUNTS_FAILURE and FETCH_CLINIC_PATIENT_COUNT_SETTINGS_FAILURE and it should call error once for a failed request', () => {
-        const clinicId = 'clinic123';
-
-        let api = {
-          clinics: {
-            getClinicPatientCount: sinon
-              .stub()
-              .callsArgWith(1, { status: 500, body: 'Count Error!' }, null),
-            getClinicPatientCountSettings: sinon
-              .stub()
-              .callsArgWith(1, { status: 500, body: 'Settings Error!' }, null),
-          },
-        };
-
-        let countErr = new Error(ErrorMessages.ERR_FETCHING_CLINIC_PATIENT_COUNTS);
-        countErr.status = 500;
-
-        let settingsErr = new Error(ErrorMessages.ERR_FETCHING_CLINIC_PATIENT_COUNT_SETTINGS);
-        settingsErr.status = 500;
-
-        let expectedActions = [
-          { type: 'SELECT_CLINIC_SUCCESS', payload: { clinicId } },
-          { type: 'tideDashboardFilters/setTideDashboardFilters', payload: undefined },
-          { type: 'FETCH_CLINIC_PATIENT_COUNTS_REQUEST' },
-          { type: 'FETCH_CLINIC_PATIENT_COUNT_SETTINGS_REQUEST' },
-          {
-            type: 'FETCH_CLINIC_PATIENT_COUNTS_FAILURE',
-            error: countErr,
-            meta: { apiError: { status: 500, body: 'Count Error!' } },
-          },
-          {
-            type: 'FETCH_CLINIC_PATIENT_COUNT_SETTINGS_FAILURE',
-            error: settingsErr,
-            meta: { apiError: { status: 500, body: 'Settings Error!' } },
-          },
-        ];
-        _.each(expectedActions, (action) => {
-          expect(isTSA(action)).to.be.true;
-        });
-
-        let store = mockStore({ blip: {
-          ...initialState,
-          clinics: {
-            [clinicId]: {
-              patientCounts: undefined,
-              patientCountSettings: undefined,
-            },
-          },
-        } });
-
-        store.dispatch(
-          async.selectClinic(api, clinicId)
-        );
-
-        const actions = store.getActions();
-
-        expect(actions[4].error).to.deep.include({
-          message: ErrorMessages.ERR_FETCHING_CLINIC_PATIENT_COUNTS,
-        });
-        expectedActions[4].error = actions[4].error;
-
-        expect(actions[5].error).to.deep.include({
-          message: ErrorMessages.ERR_FETCHING_CLINIC_PATIENT_COUNT_SETTINGS,
-        });
-        expectedActions[5].error = actions[5].error;
-        expect(actions).to.eql(expectedActions);
-
-        expect(api.clinics.getClinicPatientCount.callCount).to.equal(1);
-        expect(api.clinics.getClinicPatientCountSettings.callCount).to.equal(1);
-      });
-    });
-
     describe('fetchPatients', () => {
       let api;
       let store;
